requirements.txt
librosa
sounddevice
scipy
pillow
sounddevice
numpy
librosa
matplotlib